+++ bin/check-es-circuit-breakers.rb
@@ -88,6 +88,8 @@ class ESCircuitBreaker < Sensu::Plugin::Check::CLI
     critical 'Connection refused'
   rescue RestClient::RequestTimeout
     critical 'Connection timed out'
+  rescue RestClient::ServiceUnavailable
+    warning 'Service is unavailable'
   rescue Errno::ECONNRESET
     critical 'Connection reset by peer'
   end